+++ src/classes/dense.cpp
@@ -120,23 +120,23 @@ Dense::Dense(int64_t n_rows, int64_t n_cols)
 
 Dense::Dense(
   void (*func)(
-    double* A, uint64_t A_rows, uint64_t A_cols, uint64_t A_stride,
-    const std::vector<std::vector<double>>& x,
+    float* A, uint64_t A_rows, uint64_t A_cols, uint64_t A_stride,
+    const std::vector<std::vector<float>>& x,
     int64_t row_start, int64_t col_start
   ),
-  const std::vector<std::vector<double>>& x,
+  const std::vector<std::vector<float>>& x,
   int64_t n_rows, int64_t n_cols,
   int64_t row_start, int64_t col_start
 ) : Dense(n_rows, n_cols) {
   add_kernel_task(func, *this, x, row_start, col_start);
 }
 
-const Dense& Dense::operator=(const double a) {
+const Dense& Dense::operator=(const float a) {
   add_assign_task(*this, a);
   return *this;
 }
 
-double& Dense::operator[](int64_t i) {
+float& Dense::operator[](int64_t i) {
   assert(dim[0] == 1 || dim[1] == 1);
   if (dim[0] == 1) {
     assert(i < dim[1]);
@@ -147,7 +147,7 @@ double& Dense::operator[](int64_t i) {
   }
 }
 
-const double& Dense::operator[](int64_t i) const {
+const float& Dense::operator[](int64_t i) const {
   assert(dim[0] == 1 || dim[1] == 1);
   if (dim[0] == 1) {
     assert(i < dim[1]);
@@ -158,21 +158,21 @@ const double& Dense::operator[](int64_t i) const {
   }
 }
 
-double& Dense::operator()(int64_t i, int64_t j) {
+float& Dense::operator()(int64_t i, int64_t j) {
   assert(i < dim[0]);
   assert(j < dim[1]);
   return data_ptr[i*stride+j];
 }
 
-const double& Dense::operator()(int64_t i, int64_t j) const {
+const float& Dense::operator()(int64_t i, int64_t j) const {
   assert(i < dim[0]);
   assert(j < dim[1]);
   return data_ptr[i*stride+j];
 }
 
-double* Dense::operator&() { return data_ptr; }
+float* Dense::operator&() { return data_ptr; }
 
-const double* Dense::operator&() const { return data_ptr; }
+const float* Dense::operator&() const { return data_ptr; }
 
 Dense Dense::shallow_copy() const {
   Dense out;
