+++ src/classes/hierarchical.cpp
@@ -66,11 +66,11 @@ Hierarchical::Hierarchical(
 
 Hierarchical::Hierarchical(
   void (*func)(
-    double* A, uint64_t A_rows, uint64_t A_cols, uint64_t A_stride,
-    const std::vector<std::vector<double>>& x,
+    float* A, uint64_t A_rows, uint64_t A_cols, uint64_t A_stride,
+    const std::vector<std::vector<float>>& x,
     int64_t row_start, int64_t col_start
   ),
-  const std::vector<std::vector<double>>& x,
+  const std::vector<std::vector<float>>& x,
   int64_t n_rows, int64_t n_cols,
   int64_t rank,
   int64_t nleaf,
